add algolia indexer + remove vercel files
