--- www/apps/api-reference/vercel.json
@@ -1,10 +1,0 @@
-{
-  "crons": [{
-    "path": "/api/algolia",
-    "schedule": "0 0 * * 4"
-  }],
-  "framework": "nextjs",
-  "installCommand": "yarn install",
-  "buildCommand": "turbo run build",
-  "outputDirectory": ".next"
-}
