--- www/apps/cloud/vercel.json
@@ -1,7 +1,0 @@
-{
-  "framework": "nextjs",
-  "installCommand": "yarn install",
-  "buildCommand": "turbo run build",
-  "outputDirectory": ".next",
-  "ignoreCommand": "bash ../../ignore-build-script.sh cloud"
-}
